+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+- Fixed case in menu helper build_menu method when $item->classes is not an array (e.g. certain cases in Customizer)
+
 ## [1.21.3] - 2019-03-04
 
 ### Fixed